Custom styling
- Adjusted directory/file structure of react package. Now button and uploader reside in separate files. Moved shared types to `types.tsx`. Moved `Spinner` and function used for component rendering to `components/shared.tsx`
- Removed `ut-` prefix from state data attribute
- Adjusted `withUt` accordingly
