+++ packages/react/src/index.ts
@@ -6,5 +6,4 @@ export {
   Uploader,
   generateComponents,
 } from "./component";
-export type { UploadthingComponentProps } from "./component";
-export { withUt } from "./tw";
+export type { UploadthingComponentProps } from "./types";
