+++ packages/tw/package.json
@@ -4,9 +4,6 @@
   "description": "",
   "main": "index.js",
   "types": "types/index.d.ts",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
   "keywords": [],
   "author": "",
   "license": "ISC",
